backend/requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
bitsandbytes==0.42.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
faiss-cpu==1.9.0
fastapi==0.115.3
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==75.2.0
sniffio==1.3.1
starlette==0.41.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0